feat(commands): Add services status (table only)
Wire in the `flox services status` command. This currently only displays
table output; JSON output will be added in subsequent commits.

The bulk of the functionality is unit tested but I've also replaced all
uses of `process-compose process list` in our integration tests so that
we dogfood our own command.

This doesn't yet implement the original criteria we specified of
erroring if you provide a service that doesn't exist but I'm now less
sure whether that is necessary or desirable; interested in thoughts from
others.