feat(commands): services status --json
Implement JSON lines output for `flox services status` per:

- https://www.notion.so/floxdev/Services-status-output-53ee184e2b5b4b7d8d565041b5b555bf

Uses the existing struct and skips serialization for the fields that we
don't currently need/want to expose.

I've implemented this as an alternative display format, which seems neat
in terms of how much effort it was, but I'm not sure if it's idiomatic.