fix: check if user is logged in before instantiating environment (#1847)
Currently a `ConcreteEnvironment` is created with a `Flox` object that
may not specify auth credential to Floxhub.
While detect a missing or expired token and update the `floxhub_token`
field in `Flox`, this happens after the instantiation of the
`ConcreteEnvironment`, which remains configured with the old/absent
token. The result was that flox commands would prompt for login but fail
to finish anyway. Any subsequent run would instantiate the environment
with the now correct token.

```
$ flox rm -r limeytexan/default nodejs
You are not logged in to FloxHub. Logging in...
Go to https://auth.flox.dev/activate?user_code=ZSDT-GNMF in your browser

Your one-time activation code is: ZSDT-GNMF

✅ Authentication complete
✅ Logged in as limeytexan
❌ ERROR: Access denied to the remote environment.

This can happen if the remote is not owned by you
or the owner did not grant you access.

Please check the spelling of the remote environment
and make sure that you have access to it.

$ flox rm -r limeytexan/default nodejs
🗑️   'nodejs' uninstalled from environment 'limeytexan/default' (remote)
```

This PR changes the implementation of mutating commands (`edit`,
`install`, `uninstall`, `upgrade), to ensure a valid Floxhub token is
present before instantiating the environment, allowing the instance to
subsequently be configured correctly.

Thanks to @limeytexan for finding this.