fix: refactor to catch changes in source
If we use stdenv to pull in the source we lose the timestamps in
the process, and we rely upon those timestamps to know when to
rebuild objects in the buildCache with respect to their source.
So we instead pass in the source as a tarball to be extracted in
the sandbox at the time of the build.

This effort also brought about the realisation that I was making
the whole stable buildCache thing harder than it needed to be.
All we need to do is ensure we use the same filename across subsequent
builds, and then let Nix do the checksumming and validation to check
when things change.

Also refactored the makefile to split out template hierarchy as
follows:

  - DEPENDS_template
  - BUILD_template
    - BUILD_local_template
    - BUILD_sandbox_template

... and then selected the choice of local or sandbox template within
the loop as we iterate over $(BUILDS). Also updated a load of comments,
hopefully makes things clearer.