fix: Wait for klaus processes to finish between tests (#1875)
Klaus runs asynchronously, which means that if the last command a test
runs is `flox activate`, there is a chance it will move on to teardown
before klaus has finished its work. Specifically, klaus might try to
read and then write the env-registry.json file in the data dir, which
can fail in at least two ways:
1) Klaus itself might fail because that file has been deleted
2) The deletion of the data dir might fail because klaus has recreated
the file after it was deleted, before the directory was removed

We now provide a standard teardown step that waits for any klaus
processes associated with this test run to exit. Associated processes
are identified by the presence of the test-specific data dir in their
command line, as klaus is invoked with a reference to the env-registry
file in the data dir.