feat: virtual sandbox
The Flox "virtual sandbox" warns or aborts when encountering an ELF access
from outside the closure of packages implied by $FLOX_ENV. In this regard
it can provide the same guarantees at an ELF level provided by the sandbox
itself, but at an _advisory_ level, so that developers are informed of
missing dependencies without actually breaking anything.

The virtual sandbox is enabled by setting `FLOX_VIRTUAL_SANDBOX=(warn|abort)`
in the environment, and we do this when wrapping files in the bin directory
in the course of performing a manifest build.

As with the parsing of FLOX_ENV_LIB_DIRS, it is essential that this parsing
of the closure be performant and initialized only once per invocation, so we
start by reading closure paths into a btable from $FLOX_ENV/requisites.txt.