refactor(services): Separate status display struct
Addressing PR feedback:

- separate struct with the fields and methods for CLI display
- JSON formatting no longer overloads alternate Display
- unit test without integration with process-compose

I think it still makes sense for the types and test to live in the SDK
and for us to test the construction from `ProcessStates`.

The JSON implementation is no longer tested because I can't see a good
place to put it now that we're just using builtin serialisation.