feat(commands): Add flox services status (#1867)
## Proposed Changes

**feat(services): impl Display for ProcessStates**

Using the output format described in:

-
https://www.notion.so/floxdev/Services-status-output-53ee184e2b5b4b7d8d565041b5b555bf

I've resisted bringing in a third-party table library for the moment
because they all seem to have massive feature sets or quirks:

- https://github.com/BurntSushi/tabwriter
- https://github.com/phsym/prettytable-rs
- https://github.com/zhiburt/tabled

Some notes about the current implementation:

- NAME column is padded to 20 chars and will no longer align if it
  overflow but it will still be space delimited
- NAME is already sorted for us and the test uses header/footer entries
  to prevent any future regressions.
- PID column is right align to account variable length values

This was quick and easy to unit test with the helpers that Yannik
recently added. I opted to use real `process-compose` output, rather
than `ProcessStates` fixtures, because it's less fiddly and more
realistic than constructing all of the fields manually.

**feat(services): Add ProcessStates::filter_names**

To mutate `ProcessStates` so that it only contains the provided names.
Any unknown names are ignored. This will be used by `services status`
when providing one or more names.

**feat(commands): Add services status (table only)**

Wire in the `flox services status` command. This currently only displays
table output; JSON output will be added in subsequent commits.

The bulk of the functionality is unit tested but I've also replaced all
uses of `process-compose process list` in our integration tests so that
we dogfood our own command.

This doesn't yet implement the original criteria we specified of
erroring if you provide a service that doesn't exist but I'm now less
sure whether that is necessary or desirable; interested in thoughts from
others.

**feat(commands): services status --json**

Implement JSON lines output for `flox services status` per:

-
https://www.notion.so/floxdev/Services-status-output-53ee184e2b5b4b7d8d565041b5b555bf

Uses the existing struct and skips serialization for the fields that we
don't currently need/want to expose.

I've implemented this as an alternative display format, which seems neat
in terms of how much effort it was, but I'm not sure if it's idiomatic.

## Release Notes

Added `flox services status` command behind feature flag.