feat(commands): Indicate running PID in status out
Wrap the PID in square brackets if the process is no longer running, to
make it clear that you'll no longer be able to see the process but the
value may still be useful for cross-checking with logs.

I've exposed `is_running` from `ProcessState` but we're not currently
including it in the JSON output of `ProcessStateDisplay` as verified by
the test in the preceding commit.