fix(cuda): Adjust behaviour for nested opt-out
The test for "libcuda present but manifest opts-out" was failing for me
locally. I wasn't using a Flox environment as my default shell when I
originally wrote the test. Now that I am, the environment variable gets
inherited by the test from my default shell and doesn't get reset when a
nested activation opts-out. Fix this by always setting the value based
on the manifest of the target activation.