feat(commands): Indicate running PID in status out (#1894)
## Proposed Changes

**refactor(services): ProcessStateDisplay SDK->CLI**

Move it to the command as it's concerned with user facing display. We
have to make `ProcessState` public as a result. Added some `IntoIter`
and `From` to make `ProcessStates` easier to use from the outside.

**test(services): ProcessStateDisplay as JSON lines**

Add a test that will double check whether we expose more from the struct
than we intend to. This duplicates the iteration and `json::to_string`
in the command but it seems simple enough not to worry or warrant making
it a method on the struct right now.

**feat(commands): Indicate running PID in status out**

Wrap the PID in square brackets if the process is no longer running, to
make it clear that you'll no longer be able to see the process but the
value may still be useful for cross-checking with logs.

I've exposed `is_running` from `ProcessState` but we're not currently
including it in the JSON output of `ProcessStateDisplay` as verified by
the test in the preceding commit.

## Release Notes

Indicate inactive PIDs in `flox services status` output.