fix(cuda): Use ldconfig to detect libs faster
We've had reports of `flox activate` taking ~10s when `find` traverse
very large system library directories over slow mounts.

One option was to replace `find` with `fd` which is generally faster.
Leigh reported a difference of 2.5s to 700ms on WSL.

The other option, which was suggested by the original reporter, is to
use the `ldconfig` cache to to find the libraries. I think this is
preferable because it's super fast and we'll honour any shared library
directories that the user will have already configured on their system.

Most system packages should already trigger a regeneration of the cache
on install. Users can add `/etc/ld.so.conf.d` entries for any paths that
aren't supported by default. We no longer need to maintain a list of
known paths ourselves.

Unfortunately the diff is a little hard to read because the tests needed
refactoring to account for the new mocks and some of the existing
tests (e.g. stubbing alternate directories) no longer make sense for
this approach.