fix(cuda): Qualify (n)awk command
This wasn't causing a problem but we should include the dependency and
qualify the path so that we're not depending on the system `awk` which
may behave differently.

Using `nawk` instead of `gawk` because it's smaller (180k vs 3.2M) and
we have test coverage for the small amount of functionality that we need
from it.