feat: activation script changes needed for start (#1898)
When `flox services start name` is implemented, when `process-compose`
is not running, we will need to call `process-compose up name` in order
to start an instance of `process-compose` without starting more services
than just `name`. In this case, we'll additionally want to skip starting
the watchdog, because the activation `flox services start` is running
inside will already have a watchdog.

To support this, split the bulk of `Activate::handle` out into an
`Activate::activate` function that accepts additional arguments:
- `start_watchdog`: this will allow start and restart to skip starting a
watchdog for ephemeral activations
- `services_to_start`: `flox services start` may be called with a list
of service names, and these must be passed to the activation script

Pass `services_to_start` to the activation script as a serialized JSON
array in the `_FLOX_SERVICES_TO_START` environment variable.

This functionality will be exercised when tests are added for `flox
services start`