feat: impl `ProcessComposeLogTail` to read logs with `--tail`
Empirically, process-compose will
* log the last _up to_ `<n>` lines logged by a process
* exit with `0` exit code
* not wait for more logs of running processes, even if they logged less than `<n>` lines yet
* print a particular error message to `stderr`:

    $ process-compose process logs -u test.sock -n 2 foo 2>/dev/null
    2
    3
    write close: write unix ->test.sock: write: broken pipe

  Currently we use this as another cue that the reading completed,
  however we cannot know whether this was printed by process compose or a running service.
  A PR to banish these logs to stderr was opened upstream at
  <https://github.com/F1bonacc1/process-compose/pull/216>

This adds a new `ProcessComposeLogTail` that calls `process-compose process logs` with `--tail`
and directly reads its output, storing all lines in a vector.

Because log lines are currently not timestamped,
we can only meaningfully provide tail logs for a single process.
Thus the complexity of the multiplexed readers we need for streaming logs can be sidestepped here,
with an independent reading implementation and `Command::output`.