fix(cuda): Reintroduce fallback for NixOS
NixOS doesn't maintain a system `/etc/ld.so.cache` and generating one
doesn't appear to honour `/run/opengl-drivers` where the libraries are
commonly installed:

- https://alternativebit.fr/posts/nixos/nix-opengl-and-ubuntu-integration-nightmare/

So reintroduce that path as fallback if we don't find anything from
`ldconfig`. There shouldn't be many files or much hierachy in that
directory but we can still use `fd` as it's faster than `find`.