feat: clean up services automatically (#1871)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->

**feat: automatically shut down services**

After deregistering an environment, check if the list of activations is
now empty and if so send the shutdown command to `process-compose`.

This also redirects the watchdog stdout/stderr streams to `/dev/null`
when launched by `flox` so that none of the watchdog output appears
in the shell (especially when there's an error).

**fix: only start services when requested**

This was unconditionally starting services when the feature flag was
present.

**fix: failing test due to bad grep pattern**

**chore: add Justfile command for Nix tests**

You can now run the integration tests using Nix-built binaries via
`just nix-integ-tests` or `just test-all`.

**refactor: simplify watchdog**

This removes the async implementation in favor of a polling loop on both
operating systems. The original implementation was overly complex and
made certain errors hard to debug (e.g. a hang waiting for the signal
handler task to abort, who knows why).

This also lets us drop some dependencies and speed up compile times.

This closes #1862.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->