fix(cuda): Apply LD_AUDIT workaround for fd
This was failing silently on Linux. We have to throw away STDERR under
normal operation because we don't want anything to leak into the user
shell.

I was tempted to switch back to `find` but I'd have to re-verify it's
regex behaviour and I have no idea what the expected structure and size
of the `/run/opengl-drivers` directory is.