test(services): Remove dupe test
The retained test is a superset of the deleted test because it ensures
that no services are stopped when the invalid name is given _after_
valid names.

If we test this then we already know that it works the other way round
and it wasn't easy to tell the tell the difference between the intent of
the two tests at a glance. So we can save ourselves a bit of runtime.