test: add `logs` integration tests
* Test service name arg handling for oneshot logs and `--follow`

Oneshot logs require exactly one argument while with `--follow`
multiple logs can be specified, or omitted to default to all processes.

* Test whether without `--follow`, `logs` exits by itself
and does not wait for the service to finish or produce more logs.
More detailed tests are found in the respective unit tests for `ProcessComposeLogTail`.
Apparently `process-compose` blocks for another second or so
after printing logs almost instantaneously.
We heuristically account for that with a timeout.

* Test that --follows waits for logs and keeps receiving new lines.
Wrapping these calls in a `timeout` to ensure they are being killed
as they would otherwise run indefinitely.