fix(cuda): Use ldconfig to detect libs faster (#1902)
## Proposed Changes

We've had reports of `flox activate` taking ~10s when `find` traverse
very large system library directories over slow mounts.

One option was to replace `find` with `fd` which is generally faster.
Leigh reported a difference of 2.5s to 700ms on WSL.

The other option, which was suggested by the original reporter, is to
use the `ldconfig` cache to to find the libraries. I think this is
preferable because it's super fast and we'll honour any shared library
directories that the user will have already configured on their system.

Most system packages should already trigger a regeneration of the cache
on install. Users can add `/etc/ld.so.conf.d` entries for any paths that
aren't supported by default. We no longer need to maintain a list of
known paths ourselves.

---

NixOS doesn't maintain a system `/etc/ld.so.cache` and generating one
doesn't appear to honour `/run/opengl-drivers` where the libraries are
commonly installed:

-
https://alternativebit.fr/posts/nixos/nix-opengl-and-ubuntu-integration-nightmare/

So reintroduce that path as fallback if we don't find anything from
`ldconfig`. There shouldn't be many files or much hierachy in that
directory but we can still use `fd` as it's faster than `find`.

---

Change the fallback directory for NixOS from `/run/opengl-drivers`
(plural) to `/run/opengl-driver` (singular).

---

Unfortunately the diff is a little hard to read because the tests needed
refactoring to account for the new mocks and some of the existing tests
(e.g. stubbing alternate directories) no longer make sense for this
approach.

## Release Notes

Uses `ldconfig` to speed up detection of CUDA libraries during `flox
activate`.

You can add `ld.so.conf` entries for any additional directories that
need to be searched.

Fixed the fallback directory for drivers on NixOS.

---------

Co-authored-by: tomberek <tomberek@users.noreply.github.com>