test: avoid the use of `timeout` use pipes for synchronization
0.25 second timeouts do not reliably ensure that both services actually logged something.
The alternative implementation ensures that both processes have written _something_,
by awaiting a synchronous pipe which is written to by the processes once they wrote their first output.
The pipe will block both the test and the processes.

We keep the 0.25 second timeout to account for flox, process-compose startup and logging.
A timeout is necessary because `--follow` blocks indefinitely.