refactor: improve performance of logs by stop reading early
Calling process-compose takes about ~1 second
even though process-compose will print the logs immediately,
it will block for around a second before exiting :)
Hence we read from the output stream directly,
and break once we see the last line.
This relies on the assumption that the last line is always
the aforementioned error message,
and that this is printed to stdout, incorrectly as that may be.