ci(workflows): Add proxy builder
We're testing out a new way of running builds & tests in CI using the `--store` flag in `nix build`.
There might be some desirable side-effects this has compared to configuring the remote build machines ourselves.
One of which is that we'll just have to target 1 machine now, instead of all of them.

Co-authored-by: Bryan Honof <bryan@flox.dev>