feat: support services that spawn daemons
By default `process-compose` shuts down services by sending the service
a signal (SIGTERM by default). If a service spawns a background process
then the process that `process-compose` delivers the signal to may not
be the process that we need to stop. Other service managers, like
supervisord, get around this by having some kind of proxy process that
forwards signals to the background process. `process-compose` lets you
specify a shutdown command instead.

This change adds `is-daemon` and `shutdown.command` fields to a service
descriptor so that a user can properly configure a service that will
background itself or another process (`pg_ctl`, etc). When a user lists
an `is-daemon` service, they also need to provide a `shutdown.command`,
otherwise we have no way of shutting the process down.

Getting the `serde` machinery figured out was taking too long, so
instead of spending more time on that I simply added a `validate`
method to `ManifestServices`. In general I think we should move more
towards manual parsing/validation because we can provide better error
messages that way, but this was simply done out of expediency in this
case.