fix: re-run activation for start and restart
When `services start` and `services restart` restart `process-compose`,
they need to do so in an up-to-date activation. To support this, they
have to run the activate script. Currently the activate script skips
running hooks and setting vars if it has already been run.

Add a `_FLOX_ACTIVATE_FORCE_REACTIVATE` variable and use it to control
whether the hooks should be re-run and vars set.
This has some limitations:
- Hooks that aren't idempotent might have unexpected behavior
- Variables aren't yet unset, so if a variable is removed from an
  environment, it won't be unset

We'll address those issues later on.

Add a note to the man page saying hooks should be able to be run
multiple times non-interactively.