feat(services): Add restart command (#1911)
## Proposed Changes

This covers all of the acceptance criteria except for the following
which will come in a separate PR to speed up review:

- restarting `process-compose` to allow for reloading configs
- erroring when used outside an activation

I considered using `jq` to increment the counters in a single file but
it would have required adding another catalog fixture and also pulling
in `sponge` to update the file in-place.

Per the comment in the new manifest, I'm using `Completed` services for
most of the tests and only exercising a `Running` service when
restarting all, because the difference in time is 400ms vs 1400ms. It
didn't warrant having a completely separate manifest.

I don't appear to need to qualify the file names in the manifest because
we are using them from the current working directory in the test. This
also gets cleaned up during the test teardown.

## Release Notes

Adds `flox services restart` command behind feature flag.