feat(services): Error if no services defined
Previously `start` and `restart` would return successfully with no
output if there are no services defined in the environment.

Now all `flox services` commands return the same error and we guard
against v0 manifests in the same place.

This partially handles the cases in #1858 which were easier to solve
while I had the context from working on `start` and `restart`.

It doesn't contradict #1887 because that handles when services are
defined by they are all stopped, potentially including `process-compose`
itself.