feat(services): Indicate whether restart or start
Indicate whether `flox services restart` did a start or restart based on
whether the service was running previously.

This removes an unnecessary `ProcessStates::read` when we're using an
existing `process-compose` instance.

I considered checking the names against a HashSet of running processes
so that `ProcessStates.process()` doesn't have to filter by name each
time but passing it around made it harder to read so I've contained it
all in one private function.