fix: stop process-compose in activate
When `activate -s` is called, if process-compose is running and all
services are stopped, call `process-compose down` and then start
`process-compose` again as part of activation.

This is consistent with how `start` and `restart` handle when to restart
`process-compose`, and it allows re-starting `process-compose` in the
context of an updated environment.

The logic for starting `process-compose` in activate was refactored with
a few behavioral changes:
- We always error for remote environments if `--start-services` is
  passed. That makes it more clear services aren't ever started for
  remote environments.
- _FLOX_SERVICES_TO_START is only set when services should actually be
  started. Previously, it was dependent on the socket being present.
  Since it is only needed when services are actually started, only set
  it in that case for the sake of readability.