chore: remove `Flox` completion instance
We removed **dynamic** completions together with legacy nix wrapping commands.
The Extension trait to construct a Flox instance for us in dynamic completions was since unused.
With the update to Rust 1.80.0, clippy now complains about this being an orphan implementation.
Instead of maintaining this further, opt to remove this.
We'll need something similar once we reintroduce dynamic argument completions.