ci: Run pstree when Runner integ-tests fail
To help us figure out what's still running when the integration tests
hang. I'm only doing this on the Runner jobs, not the Nix remote builder
jobs, because it will be hard to tell on those shared instance which
processes belong to this CI run.

There's an example of this working here:

- https://github.com/flox/flox/actions/runs/10388611281/job/28764639190#step:8:1