feat(services): Move logs to .flox/log
So that they are easier for users to find and we can later garbage
collect them from a single directory based on age.

We still write a new log file for each invocation of `process-compose`
because it truncates the log file each time it starts and it will be
useful to have previous invocations to debug issues:

- https://github.com/F1bonacc1/process-compose/blob/v1.18.0/src/config/config.go#L33

I considered using the activation PID as the key in the filename and
passing the log file (rather than dir) in from `activate.rs` but I think
that there may be times when we want previous logs from a single
activation. So let `mktemp` calculate a random key for us, similar to
what we had before.

I haven't exposed the log directory as a user facing environment
variable at the moment because I don't foresee it being used dynamically
within the environment like the lib or cache directories are.

There doesn't seem a natural place for us to document the log directory,
if we're not exposing the environment variable, but it might fit into
the new docs for service management.

I'm unsure about the desired behaviour for remote environments; whether
we should mimic cache in creating temporary directories or mimic managed
environments by using `~/.cache/flox/remote/…` as we are currently.

It's worth noting that existing environments won't get the updated
`.gitignore` rules because it's only performed on `flox init`.