fix: recover from build errors due to conflicting packages
`buildenv` will fail with status 122 when realising an environment which contains multiple packages
that provide the same file.
A package can specify a `priority` to be used instead of the other to resolve the conflict.

In interactive edits, build failures of any kind were treated as unrecoverable.
However since conflict can be recovered we are returning that error as recoverable.

Thus now installing e.g. `vim` and `vim-full` to an environment with `flox edit`:

```
version = 1

[install]
vim.pkg-path = "vim"
vim-full.pkg-path = "vim-full"
```

Will prompt to continue editing:

```
❌ ERROR: 'vim' conflicts with 'vim-full'. Both packages provide the file '/bin/rview'

Resolve by uninstalling one of the conflicting packages or setting the priority of the preferred package to a value lower than '5'
! Continue editing? (Y/n)
```