feat(services): Use timestamp for services log
So that users can easily find the latest log in the directory. The
format includes microseconds in case there are multiple activations
within the same second, which is plausible when scripting commands.

I once again considered (and wrote) moving this to Rust but I think it
makes sense to keep the timestamp calculation as close to the place that
starts `process-compose` and it means that we don't have to worry about
leaking additional environment variables into the activation.