feat(services): Warn on most manifest changes
Warn the user, if they have services running for an environment, that
they will likely need to restart the services in order to pick up
manifest changes to that environment.

We originally planned to not warn about changes to the `profile` section
because those changes would only affect interactive activations. But
I've opted for a simplification for now where we treat all changes as
affecting because it would have required fragmenting the logic in
`EditResult`.

The behaviour of `install` is harder to reason about than the others.
You could have a service that depends on a binary that either isn't
available or accidentally comes from your host system which you then
want to come from the environment.

I've opted for integration tests because they were the easiest to setup
and reason about. They are grouped in `services.bats`, rather than the
individual command test files, because they share some common setup with
the other services tests.

The same changes for `pull` will come in a separate PR.