docs: service management in manifest template
Uses `postgres` as an example service.

A config file, which the user would supply themselves, is given as an
example. Opted not to provide a data directory because we haven't formed
an opinion yet about where user data should live in relation to the
environment.

The config format uses the same nesting format as packages, rather than
grouping in `[services.postgres]`.

We have to make this conditional on catalog manifests for the moment or
this integration test fails until we remove the catalog feature flag:

    ✗ activate incompatible environment fails gracefully [3865]
       tags: activate:remote:incompatible pull
       (from function `make_dummy_env' in file pull.bats, line 57,
        in test file pull.bats, line 264)
         `make_dummy_env "owner" "name"' failed
       …
       ❌ ERROR: unrecognized manifest field: 'services'.

Fixture was re-generated with:

    rm test_data/generated/envs/hello/manifest.toml
    just gen-data