fix: services logs test timing out
The test `logs: follow shows logs for all services if no names provided`
currently has a timeout of .5 seconds. We've seen that fail in CI:
https://github.com/flox/flox/actions/runs/10390132193/job/28770507812?pr=1920#step:6:480

There's likely an underlying issue, but for now, increase the timeout to
1 second. While doing so, restructure the test to background the logs
command with redirection to a file, and poll for the logs we want to
appear. That way, faster runs of the test won't wait for the full
timeout to kill the `logs --follow` command.