fix: services logs test race
The test `logs: follow will wait for logs` has flaked in CI:
https://github.com/flox/flox/actions/runs/10390132193/job/28770507812?pr=1920#step:6:425

In that failure, the 4th log line was not printed, which occurs after
the service sleeps for 3 seconds. The most likely explanation is that
process-compose was slow to start, so startup time + 3 seconds was
longer than the 4 second timeout we have until failing the test.

Add a fifo to ensure that the 4 second timeout doesn't start until
after process-compose has started the service and the service has
printed the first few logs.