feat(services): Warn on flox pull changes
Add the same warning to `flox pull` as we have for `install`,
`uninstall`, `upgrade`, and `edit`.

This only covers the happy path where you are successfully pulling into
an existing environment. In the case where a pull fails and needs to be
manually resolved the warning would be emitted on one of the other
commands.

We don't currently emit any context specific warnings if you forcibly
pull over the top of an existing environment. I think that's OK for the
moment though because:

- it's inline with not warning about existing activations and not warning on `flox delete`
- services will still be cleaned up by the watchdog when all existing activations exit