chore: handle service connection errors
Add more information to connection errors when talking to the service manager.

Process-Compose commands may fail with [LoggedError::SocketDoesntExist]
if the specific socket doesn't exist,
or if the service manager has quit without cleaning up the socket,
i.e. the socket is unresponsive.

This function adds more context to the error in the latter case.

For practical purposes, we apply this to the [ProcessStates::read] call in
all services commands, as that is typically the first call that can fail.
For following commands, we can assume that the service manager is running.