chore: wrap checked `ConcreteEnvironment`
* Add a `ServicesEnvironment` type that wraps a `ConcreteEnvironment`
  that has been checked for compatibility with service commands.

Constructing a `ServicesEnvironment` requires a local `ConcreteEnvironment`
with a v1 manifest.

* implement "guard" methods for `ServicesEnvironment` to check against common command invariants

To use service commands, we require that the service manager socket exists
and that there are services defined in the environment at all.
The environment's `socket` is used to determine whether services are running,
which may not be the most robust solution, but is currently used consistently.

`start` and `restart` additionally require the current process to run within an activation of the environment.