refactor: parse process-compose socket connect error
Instead of trying to guard commands by the presence of a socket file
and a singular call to `status`, extend the process-compose error parser
to handle "connection refused" error messages.

Additionally add an error pretty printer for human domain error messages,
when the socket file is missing or connecting to it failed.