chore: delete unused pkgdb functionality
The only change to the Rust side is removing the body of the update
command. This also moves one test from `flox-nixpkgs.bats` to
`flox-nixpkgs.cc`. This test relied on the `pkgdb get` command. Moving
the test to C++ allows us to drop `pkgdb get` entirely.

The only other potential change to behavior is that the
`primop_getFingerprint` function relied on the `FloxFlake` abstraction,
but we dropped that because it's not really needed. Instead we're now
locking using Nix functions and using the `defaultLockFlags` that
previously existed.