fix(tracing): Change fields to tags
Replace all uses of `instrument` and `record` to set fields with a
helper function that sets tags on the current scope. This allows us to
search them in the Sentry UI.

Tags are the only custom keys that you can search on:

- https://docs.sentry.io/concepts/search/searchable-properties/
- https://docs.sentry.io/platforms/rust/enriching-events/tags/
- https://docs.sentry.io/concepts/key-terms/enrich-data/#types-of-data
- https://sentry.zendesk.com/hc/en-us/articles/26861002545691-How-to-search-context

The other-other form of sending arbitrary data has been deprecated:

- https://docs.sentry.io/platforms/rust/enriching-events/context/#additional-data

They have no cost implications because billing is by span quantity:

- https://sentry.io/pricing/

The docstring explains more about why we have to do this currently. I
attempted automating the conversion in-house but the following only have
access to events and not transactions/spans:

- `ClientOptions.integrations`
- `ClientOptions.before_send`